+++ ocrd/cli/workspace.py
@@ -118,6 +118,35 @@ def workspace_add_file(ctx, file_grp, file_id, mimetype, group_id, local_filenam
     workspace.save_mets()
 
 # ----------------------------------------------------------------------
+# ocrd workspace find
+# ----------------------------------------------------------------------
+
+@workspace_cli.command('find', help="""
+
+    Find files
+
+""")
+@click.option('-G', '--file-grp', help="fileGrp USE")
+@click.option('-m', '--mimetype', help="Media type to look for")
+@click.option('-g', '--group-id', help="GROUPID")
+@click.option('-l', '--local-filename', is_flag=True, help="Whether to print the local_filename of files")
+@click.option('-a', '--download-all', is_flag=True, help="Whether to download found files")
+@pass_workspace
+def workspace_find(ctx, file_grp, mimetype, group_id, local_filename, download_all):
+    workspace = Workspace(ctx.resolver, directory=ctx.directory)
+    for f in workspace.mets.find_files(
+            fileGrp=file_grp,
+            mimetype=mimetype,
+            groupId=group_id,
+        ):
+        if download_all:
+            workspace.download_file(f)
+        if local_filename:
+            print(f.local_filename)
+        else:
+            print(f.url)
+
+# ----------------------------------------------------------------------
 # ocrd workspace pack
 # ----------------------------------------------------------------------